Refactor EncounterList into separate component
